requirements.txt
boto3==1.9.164
GitPython==3.1.0
PyYAML==5.1.1
py-cid==0.2.1
py-multihash==0.2.3
botocore==1.12.172
colorama==0.4.1
tqdm==4.31.1
schema==0.7.0
click==7.0
click-didyoumean==0.0.3
click-plugins==1.1.1
hurry.filesize==0.9
azure-core==1.3.0
azure-storage-blob==12.0.0
toml==0.10.0